pyproject.toml
[project]
name = "agent-mcp-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.39.0",
    "google-api-python-client>=2.167.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.0.0",
]